pyproject.toml
[tool.poetry]
name = "parma-bot"
version = "0.1.0"
description = ""
authors = ["gas <pankot222@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
